+++ src/features/island/hud/components/codex/Codex.tsx
@@ -25,13 +25,10 @@ import { useSound } from "lib/utils/hooks/useSound";
 import trophy from "assets/icons/trophy.png";
 import factions from "assets/icons/factions.webp";
 import chores from "assets/icons/chores.webp";
-import factionMark from "assets/icons/faction_mark.webp";
 import { TicketsLeaderboard } from "./pages/TicketsLeaderboard";
 import { Leaderboards } from "features/game/expansion/components/leaderboard/actions/cache";
 import { fetchLeaderboardData } from "features/game/expansion/components/leaderboard/actions/leaderboard";
-import { hasFeatureAccess } from "lib/flags";
-import { FactionsLeaderboard } from "./pages/FactionsLeaderboard";
-import { MarksLeaderboard } from "./pages/MarksLeaderboard";
+import { FactionLeaderboard } from "./pages/FactionLeaderboard";
 
 interface Props {
   show: boolean;
@@ -139,17 +136,8 @@ export const Codex: React.FC<Props> = ({ show, onHide }) => {
     ...(state.faction
       ? [
           {
-            name: "Factions" as const,
-            icon: factions,
-            count: 0,
-          },
-        ]
-      : []),
-    ...(hasFeatureAccess(state, "MARKS_LEADERBOARD")
-      ? [
-          {
             name: "Marks" as const,
-            icon: factionMark,
+            icon: factions,
             count: 0,
           },
         ]
@@ -243,25 +231,10 @@ export const Codex: React.FC<Props> = ({ show, onHide }) => {
                 />
               </InnerPanel>
             )}
+
             {currentTab === 5 && state.faction && (
-              <InnerPanel
-                className={classNames(
-                  "flex flex-col h-full overflow-hidden overflow-y-auto scrollable",
-                )}
-              >
-                <FactionsLeaderboard
-                  id={id}
-                  faction={state.faction.name}
-                  isLoading={data?.emblems === undefined}
-                  data={data?.emblems?.emblems ?? null}
-                  lastUpdated={data?.kingdom?.lastUpdated ?? null}
-                />
-              </InnerPanel>
-            )}
-            {currentTab === 6 && state.faction && (
-              <MarksLeaderboard
-                emblemLeaderboard={data?.emblems ?? null}
-                marksLeaderboard={data?.kingdom ?? null}
+              <FactionLeaderboard
+                leaderboard={data?.kingdom ?? null}
                 isLoading={data?.kingdom === undefined}
                 playerId={id}
                 faction={state.faction.name}
