+++ src/lib/flags.ts
@@ -36,7 +36,6 @@ export type FeatureName =
   | "DESERT_RECIPES"
   | "FACTION_HOUSE"
   | "CROP_QUICK_SELECT"
-  | "MARKS_LEADERBOARD"
   | "FESTIVAL_OF_COLORS"
   | "FACTION_KITCHEN"
   | "FACTION_CHORES"
@@ -67,7 +66,6 @@ const featureFlags: Record<FeatureName, FeatureFlag> = {
   BANNER_SALES: clashOfFactionsFeatureFlag,
   // Just in case we need to disable the crop machine, leave the flag in temporarily
   CROP_MACHINE: () => true,
-  MARKS_LEADERBOARD: betaTimeBasedFeatureFlag(new Date("2022-07-08T00:00:00Z")),
   FACTION_KITCHEN: betaTimeBasedFeatureFlag(new Date("2022-07-08T00:00:00Z")),
   FACTION_CHORES: betaTimeBasedFeatureFlag(new Date("2022-07-08T00:00:00Z")),
 };
